Strip strings before calling .empty?
Currently, passing empty strings as options, such as ' ', or '    '.  generates an invalid dhcpd.conf file. The foreman_setup plugin asks you to run `--foreman-proxy-dhcp-range = "  "` if the range is unspecified in the UI. Then the installer or the user won't be able to start the service unless they manually dig into the dhcpd.conf file.

This could happen with any of the options so that's why I changed it everywhere.