add $ddns_updates variable to init.pp
update dhcpd.conf.erb set ddns to on if the $ddns_updates variable is set

This is to allow the module to turn ddns updates on even if secure updates via a key are not enalbled. I know... shame shame.
BUT this is helpful when using windows based dns servers along with isc-dhcp-server

update erb template

revert my previous change

allow insecure dns updates when no key specified

add $ddns_updates variable to init.pp
update dhcpd.conf.erb set ddns to on if the $ddns_updates variable is set

This is to allow the module to turn ddns updates on even if secure updates via a key are not enalbled. I know... shame shame.
BUT this is helpful when using windows based dns servers along with isc-dhcp-server

remove extra end tag in dhcp.conf.erb

fix and and ors to follow ruby convention

add () to properly contain || expression

add new line to end of file