allow insecure dns updates when no key specified
add $ddns_updates variable to init.pp
update dhcpd.conf.erb set ddns to on if the $ddns_updates variable is set

This is to allow the module to turn ddns updates on even if secure updates via a key are not enalbled. I know... shame shame.
BUT this is helpful when using windows based dns servers along with isc-dhcp-server